Updated DeleteDeviceSessionUseCase

+++ src/features/usecases/deleteDeviceSessionUseCase.ts
@@ -1,7 +1,5 @@
-import {BlogsRepository} from "../blogs/infrastructure/blogs.repository";
 import {CommandHandler, ICommandHandler} from "@nestjs/cqrs";
-import {BlogsQueryRepository} from "../blogs/infrastructure/blogs.query-repository";
-import {NotFoundException} from "@nestjs/common";
+import {ForbiddenException, NotFoundException} from "@nestjs/common";
 import {SessionRepository} from "../session/infrastructure/session.repository";
 import {SessionQueryRepository} from "../session/infrastructure/session.query-repository";
 
@@ -22,13 +20,18 @@ export class DeleteDeviceSessionUseCase implements ICommandHandler<DeleteDeviceS
 
     async execute(command: DeleteDeviceSessionUseCaseCommand) {
 
-        const session = await this.sessionQueryRepository.getUserSession(command.userId, command.deviceId);
+        const session = await this.sessionQueryRepository.getUserSessionByDeviceId(command.deviceId);
 
         if (!session) {
             throw new NotFoundException('Session not found');
         }
 
+        if (command.userId !== session.userId) {
+            throw new ForbiddenException('Session deletion denied');
+        }
+
         return await this.sessionRepository.deleteSession(command.userId, command.deviceId);
+
     }
 
 }
